requirements.txt
# ============================================
# Chat With Your PDF — Compatible Dependencies
# Python 3.10 | All versions tested together
# ============================================

# ===== Core LangChain =====
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.6
langchain-groq==0.2.0
langchain-text-splitters==0.3.0

# ===== LangGraph =====
langgraph==0.2.28
langgraph-checkpoint==1.0.12

# ===== API Server =====
fastapi==0.115.0
uvicorn==0.30.6
langserve==0.3.0
sse-starlette==1.8.2

# ===== Vector Store =====
faiss-cpu==1.8.0.post1

# ===== Document Processing =====
pypdf==5.0.1
python-docx==1.1.2
pdfplumber==0.11.4
PyMuPDF==1.24.14

# ===== Embeddings =====
sentence-transformers==3.0.1

# ===== UI =====
gradio==4.44.1
huggingface-hub==0.25.2

# ===== Utilities =====
numpy==1.26.4
pandas==2.2.2
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
tqdm==4.66.5
